V3.8.0 pipeline (#15428)
