+++ cocos/rendering/custom/types.ts
@@ -28,10 +28,10 @@
  * ========================= !DO NOT CHANGE THE FOLLOWING SECTION MANUALLY! =========================
  */
 /* eslint-disable max-len */
-import { ClearFlagBit, Color, LoadOp, ResolveMode, ShaderStageFlagBit, StoreOp, Type, UniformBlock } from '../../gfx';
+import { ResolveMode, ShaderStageFlagBit, Type, UniformBlock } from '../../gfx';
 import { Light } from '../../render-scene/scene';
 import { OutputArchive, InputArchive } from './archive';
-import { saveColor, loadColor, saveUniformBlock, loadUniformBlock } from './serialization';
+import { saveUniformBlock, loadUniformBlock } from './serialization';
 
 export enum UpdateFrequency {
     PER_INSTANCE,
@@ -262,38 +262,6 @@ export function getAccessTypeName (e: AccessType): string {
     }
 }
 
-export class RasterView {
-    constructor (
-        slotName = '',
-        accessType: AccessType = AccessType.WRITE,
-        attachmentType: AttachmentType = AttachmentType.RENDER_TARGET,
-        loadOp: LoadOp = LoadOp.LOAD,
-        storeOp: StoreOp = StoreOp.STORE,
-        clearFlags: ClearFlagBit = ClearFlagBit.ALL,
-        clearColor: Color = new Color(),
-        shaderStageFlags: ShaderStageFlagBit = ShaderStageFlagBit.NONE,
-    ) {
-        this.slotName = slotName;
-        this.accessType = accessType;
-        this.attachmentType = attachmentType;
-        this.loadOp = loadOp;
-        this.storeOp = storeOp;
-        this.clearFlags = clearFlags;
-        this.clearColor = clearColor;
-        this.shaderStageFlags = shaderStageFlags;
-    }
-    slotName: string;
-    slotName1 = '';
-    accessType: AccessType;
-    attachmentType: AttachmentType;
-    loadOp: LoadOp;
-    storeOp: StoreOp;
-    clearFlags: ClearFlagBit;
-    readonly clearColor: Color;
-    slotID = 0;
-    shaderStageFlags: ShaderStageFlagBit;
-}
-
 export enum ClearValueType {
     NONE,
     FLOAT_TYPE,
@@ -311,44 +279,6 @@ export function getClearValueTypeName (e: ClearValueType): string {
     default:
         return '';
     }
-}
-
-export class ClearValue {
-    constructor (x = 0, y = 0, z = 0, w = 0) {
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.w = w;
-    }
-    x: number;
-    y: number;
-    z: number;
-    w: number;
-}
-
-export class ComputeView {
-    constructor (
-        name = '',
-        accessType: AccessType = AccessType.READ,
-        clearFlags: ClearFlagBit = ClearFlagBit.NONE,
-        clearValueType: ClearValueType = ClearValueType.NONE,
-        clearValue: ClearValue = new ClearValue(),
-        shaderStageFlags: ShaderStageFlagBit = ShaderStageFlagBit.NONE,
-    ) {
-        this.name = name;
-        this.accessType = accessType;
-        this.clearFlags = clearFlags;
-        this.clearValueType = clearValueType;
-        this.clearValue = clearValue;
-        this.shaderStageFlags = shaderStageFlags;
-    }
-    name: string;
-    accessType: AccessType;
-    plane = 0;
-    clearFlags: ClearFlagBit;
-    clearValueType: ClearValueType;
-    readonly clearValue: ClearValue;
-    shaderStageFlags: ShaderStageFlagBit;
 }
 
 export class LightInfo {
@@ -564,66 +494,6 @@ export class PipelineStatistics {
     numFreeDescriptorSets = 0;
     numInstancingBuffers = 0;
     numInstancingUniformBlocks = 0;
-}
-
-export function saveRasterView (ar: OutputArchive, v: RasterView) {
-    ar.writeString(v.slotName);
-    ar.writeString(v.slotName1);
-    ar.writeNumber(v.accessType);
-    ar.writeNumber(v.attachmentType);
-    ar.writeNumber(v.loadOp);
-    ar.writeNumber(v.storeOp);
-    ar.writeNumber(v.clearFlags);
-    saveColor(ar, v.clearColor);
-    ar.writeNumber(v.slotID);
-    ar.writeNumber(v.shaderStageFlags);
-}
-
-export function loadRasterView (ar: InputArchive, v: RasterView) {
-    v.slotName = ar.readString();
-    v.slotName1 = ar.readString();
-    v.accessType = ar.readNumber();
-    v.attachmentType = ar.readNumber();
-    v.loadOp = ar.readNumber();
-    v.storeOp = ar.readNumber();
-    v.clearFlags = ar.readNumber();
-    loadColor(ar, v.clearColor);
-    v.slotID = ar.readNumber();
-    v.shaderStageFlags = ar.readNumber();
-}
-
-export function saveClearValue (ar: OutputArchive, v: ClearValue) {
-    ar.writeNumber(v.x);
-    ar.writeNumber(v.y);
-    ar.writeNumber(v.z);
-    ar.writeNumber(v.w);
-}
-
-export function loadClearValue (ar: InputArchive, v: ClearValue) {
-    v.x = ar.readNumber();
-    v.y = ar.readNumber();
-    v.z = ar.readNumber();
-    v.w = ar.readNumber();
-}
-
-export function saveComputeView (ar: OutputArchive, v: ComputeView) {
-    ar.writeString(v.name);
-    ar.writeNumber(v.accessType);
-    ar.writeNumber(v.plane);
-    ar.writeNumber(v.clearFlags);
-    ar.writeNumber(v.clearValueType);
-    saveClearValue(ar, v.clearValue);
-    ar.writeNumber(v.shaderStageFlags);
-}
-
-export function loadComputeView (ar: InputArchive, v: ComputeView) {
-    v.name = ar.readString();
-    v.accessType = ar.readNumber();
-    v.plane = ar.readNumber();
-    v.clearFlags = ar.readNumber();
-    v.clearValueType = ar.readNumber();
-    loadClearValue(ar, v.clearValue);
-    v.shaderStageFlags = ar.readNumber();
 }
 
 export function saveLightInfo (ar: OutputArchive, v: LightInfo) {
