+++ native/cocos/renderer/pipeline/custom/NativePipelineTypes.h
@@ -107,7 +107,6 @@ public:
     void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) /*implements*/;
     void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) /*implements*/;
     void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) /*implements*/;
-    void addComputeView(const ccstd::string &name, const ComputeView &view) /*implements*/;
     void setViewport(const gfx::Viewport &viewport) /*implements*/;
     RenderQueueBuilder *addQueue(QueueHint hint, const ccstd::string &phaseName) /*implements*/;
     bool getShowStatistics() const /*implements*/;
@@ -242,9 +241,6 @@ public:
     void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override {
         NativeRenderSubpassBuilderImpl::addStorageImage(name, accessType, slotName);
     }
-    void addComputeView(const ccstd::string &name, const ComputeView &view) override {
-        NativeRenderSubpassBuilderImpl::addComputeView(name, view);
-    }
     void setViewport(const gfx::Viewport &viewport) override {
         NativeRenderSubpassBuilderImpl::setViewport(viewport);
     }
@@ -329,9 +325,6 @@ public:
     void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override {
         NativeRenderSubpassBuilderImpl::addStorageImage(name, accessType, slotName);
     }
-    void addComputeView(const ccstd::string &name, const ComputeView &view) override {
-        NativeRenderSubpassBuilderImpl::addComputeView(name, view);
-    }
     void setViewport(const gfx::Viewport &viewport) override {
         NativeRenderSubpassBuilderImpl::setViewport(viewport);
     }
@@ -408,7 +401,6 @@ public:
     void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) override;
     void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override;
     void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override;
-    void addComputeView(const ccstd::string &name, const ComputeView &view) override;
     ComputeQueueBuilder *addQueue(const ccstd::string &phaseName) override;
     void setCustomShaderStages(const ccstd::string &name, gfx::ShaderStageFlagBit stageFlags) override;
 };
@@ -468,8 +460,6 @@ public:
     void addRenderTarget(const ccstd::string &name, gfx::LoadOp loadOp, gfx::StoreOp storeOp, const gfx::Color &color) override;
     void addDepthStencil(const ccstd::string &name, gfx::LoadOp loadOp, gfx::StoreOp storeOp, float depth, uint8_t stencil, gfx::ClearFlagBit clearFlags) override;
     void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) override;
-    void addRasterView(const ccstd::string &name, const RasterView &view) override;
-    void addComputeView(const ccstd::string &name, const ComputeView &view) override;
     RenderQueueBuilder *addQueue(QueueHint hint, const ccstd::string &phaseName) override;
     void setViewport(const gfx::Viewport &viewport) override;
     void setVersion(const ccstd::string &name, uint64_t version) override;
@@ -596,7 +586,6 @@ public:
     void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override;
     void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) override;
     void addMaterialTexture(const ccstd::string &resourceName, gfx::ShaderStageFlagBit flags) override;
-    void addComputeView(const ccstd::string &name, const ComputeView &view) override;
     ComputeQueueBuilder *addQueue(const ccstd::string &phaseName) override;
     void setCustomShaderStages(const ccstd::string &name, gfx::ShaderStageFlagBit stageFlags) override;
 };
