+++ native/cocos/renderer/pipeline/custom/NativeProgramLibrary.cpp
@@ -1265,7 +1265,7 @@ void NativeProgramLibrary::addEffect(const EffectAsset *effectAssetIn) {
             const auto &srcShaderInfo = *pShaderInfo;
             CC_ENSURES(subpassOrPassID != INVALID_ID && phaseID != INVALID_ID);
             const auto &passLayout = get(LayoutGraphData::LayoutTag{}, lg, subpassOrPassID);
-            const auto &phaseLayout = get(LayoutGraphData::LayoutTag{}, lg, subpassOrPassID);
+            const auto &phaseLayout = get(LayoutGraphData::LayoutTag{}, lg, phaseID);
 
             // programs
             auto iter = this->phases.find(phaseID);
