+++ native/cocos/renderer/pipeline/custom/RenderCommonJsb.cpp
@@ -34,93 +34,6 @@
 #include "cocos/renderer/pipeline/custom/RenderCommonTypes.h"
 #include "cocos/renderer/pipeline/custom/details/JsbConversion.h"
 
-bool nativevalue_to_se(const cc::render::RasterView &from, se::Value &to, se::Object *ctx) { // NOLINT
-    se::HandleObject obj(se::Object::createPlainObject());
-    se::Value        tmp;
-
-    nativevalue_to_se(from.slotName, tmp, ctx);
-    obj->setProperty("slotName", tmp);
-
-    nativevalue_to_se(from.slotName1, tmp, ctx);
-    obj->setProperty("slotName1", tmp);
-
-    nativevalue_to_se(from.accessType, tmp, ctx);
-    obj->setProperty("accessType", tmp);
-
-    nativevalue_to_se(from.attachmentType, tmp, ctx);
-    obj->setProperty("attachmentType", tmp);
-
-    nativevalue_to_se(from.loadOp, tmp, ctx);
-    obj->setProperty("loadOp", tmp);
-
-    nativevalue_to_se(from.storeOp, tmp, ctx);
-    obj->setProperty("storeOp", tmp);
-
-    nativevalue_to_se(from.clearFlags, tmp, ctx);
-    obj->setProperty("clearFlags", tmp);
-
-    nativevalue_to_se(from.clearColor, tmp, ctx);
-    obj->setProperty("clearColor", tmp);
-
-    nativevalue_to_se(from.slotID, tmp, ctx);
-    obj->setProperty("slotID", tmp);
-
-    nativevalue_to_se(from.shaderStageFlags, tmp, ctx);
-    obj->setProperty("shaderStageFlags", tmp);
-
-    to.setObject(obj);
-    return true;
-}
-
-bool nativevalue_to_se(const cc::render::ClearValue &from, se::Value &to, se::Object *ctx) { // NOLINT
-    se::HandleObject obj(se::Object::createPlainObject());
-    se::Value        tmp;
-
-    nativevalue_to_se(from.x, tmp, ctx);
-    obj->setProperty("x", tmp);
-
-    nativevalue_to_se(from.y, tmp, ctx);
-    obj->setProperty("y", tmp);
-
-    nativevalue_to_se(from.z, tmp, ctx);
-    obj->setProperty("z", tmp);
-
-    nativevalue_to_se(from.w, tmp, ctx);
-    obj->setProperty("w", tmp);
-
-    to.setObject(obj);
-    return true;
-}
-
-bool nativevalue_to_se(const cc::render::ComputeView &from, se::Value &to, se::Object *ctx) { // NOLINT
-    se::HandleObject obj(se::Object::createPlainObject());
-    se::Value        tmp;
-
-    nativevalue_to_se(from.name, tmp, ctx);
-    obj->setProperty("name", tmp);
-
-    nativevalue_to_se(from.accessType, tmp, ctx);
-    obj->setProperty("accessType", tmp);
-
-    nativevalue_to_se(from.plane, tmp, ctx);
-    obj->setProperty("plane", tmp);
-
-    nativevalue_to_se(from.clearFlags, tmp, ctx);
-    obj->setProperty("clearFlags", tmp);
-
-    nativevalue_to_se(from.clearValueType, tmp, ctx);
-    obj->setProperty("clearValueType", tmp);
-
-    nativevalue_to_se(from.clearValue, tmp, ctx);
-    obj->setProperty("clearValue", tmp);
-
-    nativevalue_to_se(from.shaderStageFlags, tmp, ctx);
-    obj->setProperty("shaderStageFlags", tmp);
-
-    to.setObject(obj);
-    return true;
-}
-
 bool nativevalue_to_se(const cc::render::LightInfo &from, se::Value &to, se::Object *ctx) { // NOLINT
     se::HandleObject obj(se::Object::createPlainObject());
     se::Value        tmp;
@@ -312,120 +225,6 @@ bool nativevalue_to_se(const cc::render::MovePair &from, se::Value &to, se::Obje
 
     to.setObject(obj);
     return true;
-}
-
-template <>
-bool sevalue_to_native<cc::render::RasterView>(const se::Value &from, cc::render::RasterView *to, se::Object *ctx) { // NOLINT
-    SE_PRECONDITION2(from.isObject(), false, " Convert parameter to RasterView failed !");
-
-    auto *obj = const_cast<se::Object *>(from.toObject());
-    bool ok = true;
-    se::Value field;
-    obj->getProperty("slotName", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->slotName), ctx);
-    }
-    obj->getProperty("slotName1", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->slotName1), ctx);
-    }
-    obj->getProperty("accessType", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->accessType), ctx);
-    }
-    obj->getProperty("attachmentType", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->attachmentType), ctx);
-    }
-    obj->getProperty("loadOp", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->loadOp), ctx);
-    }
-    obj->getProperty("storeOp", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->storeOp), ctx);
-    }
-    obj->getProperty("clearFlags", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->clearFlags), ctx);
-    }
-    obj->getProperty("clearColor", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->clearColor), ctx);
-    }
-    obj->getProperty("slotID", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->slotID), ctx);
-    }
-    obj->getProperty("shaderStageFlags", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->shaderStageFlags), ctx);
-    }
-    return ok;
-}
-
-template <>
-bool sevalue_to_native<cc::render::ClearValue>(const se::Value &from, cc::render::ClearValue *to, se::Object *ctx) { // NOLINT
-    SE_PRECONDITION2(from.isObject(), false, " Convert parameter to ClearValue failed !");
-
-    auto *obj = const_cast<se::Object *>(from.toObject());
-    bool ok = true;
-    se::Value field;
-    obj->getProperty("x", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->x), ctx);
-    }
-    obj->getProperty("y", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->y), ctx);
-    }
-    obj->getProperty("z", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->z), ctx);
-    }
-    obj->getProperty("w", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->w), ctx);
-    }
-    return ok;
-}
-
-template <>
-bool sevalue_to_native<cc::render::ComputeView>(const se::Value &from, cc::render::ComputeView *to, se::Object *ctx) { // NOLINT
-    SE_PRECONDITION2(from.isObject(), false, " Convert parameter to ComputeView failed !");
-
-    auto *obj = const_cast<se::Object *>(from.toObject());
-    bool ok = true;
-    se::Value field;
-    obj->getProperty("name", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->name), ctx);
-    }
-    obj->getProperty("accessType", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->accessType), ctx);
-    }
-    obj->getProperty("plane", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->plane), ctx);
-    }
-    obj->getProperty("clearFlags", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->clearFlags), ctx);
-    }
-    obj->getProperty("clearValueType", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->clearValueType), ctx);
-    }
-    obj->getProperty("clearValue", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->clearValue), ctx);
-    }
-    obj->getProperty("shaderStageFlags", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->shaderStageFlags), ctx);
-    }
-    return ok;
 }
 
 template <>
