+++ native/cocos/renderer/pipeline/custom/RenderCommonJsb.h
@@ -32,12 +32,6 @@
 #include "cocos/bindings/manual/jsb_conversions.h"
 #include "cocos/renderer/pipeline/custom/RenderCommonFwd.h"
 
-bool nativevalue_to_se(const cc::render::RasterView &from, se::Value &to, se::Object *ctx); // NOLINT
-
-bool nativevalue_to_se(const cc::render::ClearValue &from, se::Value &to, se::Object *ctx); // NOLINT
-
-bool nativevalue_to_se(const cc::render::ComputeView &from, se::Value &to, se::Object *ctx); // NOLINT
-
 bool nativevalue_to_se(const cc::render::LightInfo &from, se::Value &to, se::Object *ctx); // NOLINT
 
 bool nativevalue_to_se(const cc::render::Descriptor &from, se::Value &to, se::Object *ctx); // NOLINT
@@ -55,15 +49,6 @@ bool nativevalue_to_se(const cc::render::UploadPair &from, se::Value &to, se::Ob
 bool nativevalue_to_se(const cc::render::MovePair &from, se::Value &to, se::Object *ctx); // NOLINT
 
 // if function overload is used, android build fails
-template <>
-bool sevalue_to_native(const se::Value &from, cc::render::RasterView *to, se::Object *ctx); // NOLINT
-
-template <>
-bool sevalue_to_native(const se::Value &from, cc::render::ClearValue *to, se::Object *ctx); // NOLINT
-
-template <>
-bool sevalue_to_native(const se::Value &from, cc::render::ComputeView *to, se::Object *ctx); // NOLINT
-
 template <>
 bool sevalue_to_native(const se::Value &from, cc::render::LightInfo *to, se::Object *ctx); // NOLINT
