+++ native/cocos/renderer/pipeline/custom/RenderCommonSerialization.h
@@ -37,66 +37,6 @@ namespace cc {
 
 namespace render {
 
-inline void save(OutputArchive& ar, const RasterView& v) {
-    save(ar, v.slotName);
-    save(ar, v.slotName1);
-    save(ar, v.accessType);
-    save(ar, v.attachmentType);
-    save(ar, v.loadOp);
-    save(ar, v.storeOp);
-    save(ar, v.clearFlags);
-    save(ar, v.clearColor);
-    save(ar, v.slotID);
-    save(ar, v.shaderStageFlags);
-}
-
-inline void load(InputArchive& ar, RasterView& v) {
-    load(ar, v.slotName);
-    load(ar, v.slotName1);
-    load(ar, v.accessType);
-    load(ar, v.attachmentType);
-    load(ar, v.loadOp);
-    load(ar, v.storeOp);
-    load(ar, v.clearFlags);
-    load(ar, v.clearColor);
-    load(ar, v.slotID);
-    load(ar, v.shaderStageFlags);
-}
-
-inline void save(OutputArchive& ar, const ClearValue& v) {
-    save(ar, v.x);
-    save(ar, v.y);
-    save(ar, v.z);
-    save(ar, v.w);
-}
-
-inline void load(InputArchive& ar, ClearValue& v) {
-    load(ar, v.x);
-    load(ar, v.y);
-    load(ar, v.z);
-    load(ar, v.w);
-}
-
-inline void save(OutputArchive& ar, const ComputeView& v) {
-    save(ar, v.name);
-    save(ar, v.accessType);
-    save(ar, v.plane);
-    save(ar, v.clearFlags);
-    save(ar, v.clearValueType);
-    save(ar, v.clearValue);
-    save(ar, v.shaderStageFlags);
-}
-
-inline void load(InputArchive& ar, ComputeView& v) {
-    load(ar, v.name);
-    load(ar, v.accessType);
-    load(ar, v.plane);
-    load(ar, v.clearFlags);
-    load(ar, v.clearValueType);
-    load(ar, v.clearValue);
-    load(ar, v.shaderStageFlags);
-}
-
 inline void save(OutputArchive& ar, const LightInfo& v) {
     // skip, light: IntrusivePtr<scene::Light>
     save(ar, v.level);
