+++ native/cocos/renderer/pipeline/custom/RenderCommonTypes.cpp
@@ -34,94 +34,6 @@ namespace cc {
 
 namespace render {
 
-RasterView::RasterView(const allocator_type& alloc) noexcept
-: slotName(alloc),
-  slotName1(alloc) {}
-
-RasterView::RasterView(ccstd::pmr::string slotNameIn, AccessType accessTypeIn, AttachmentType attachmentTypeIn, gfx::LoadOp loadOpIn, gfx::StoreOp storeOpIn, gfx::ClearFlagBit clearFlagsIn, gfx::Color clearColorIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc) noexcept // NOLINT
-: slotName(std::move(slotNameIn), alloc),
-  slotName1(alloc),
-  accessType(accessTypeIn),
-  attachmentType(attachmentTypeIn),
-  loadOp(loadOpIn),
-  storeOp(storeOpIn),
-  clearFlags(clearFlagsIn),
-  clearColor(clearColorIn),
-  shaderStageFlags(shaderStageFlagsIn) {}
-
-RasterView::RasterView(ccstd::pmr::string slotNameIn, ccstd::pmr::string slotName1In, AccessType accessTypeIn, AttachmentType attachmentTypeIn, gfx::LoadOp loadOpIn, gfx::StoreOp storeOpIn, gfx::ClearFlagBit clearFlagsIn, gfx::Color clearColorIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc) noexcept // NOLINT
-: slotName(std::move(slotNameIn), alloc),
-  slotName1(std::move(slotName1In), alloc),
-  accessType(accessTypeIn),
-  attachmentType(attachmentTypeIn),
-  loadOp(loadOpIn),
-  storeOp(storeOpIn),
-  clearFlags(clearFlagsIn),
-  clearColor(clearColorIn),
-  shaderStageFlags(shaderStageFlagsIn) {}
-
-RasterView::RasterView(RasterView&& rhs, const allocator_type& alloc)
-: slotName(std::move(rhs.slotName), alloc),
-  slotName1(std::move(rhs.slotName1), alloc),
-  accessType(rhs.accessType),
-  attachmentType(rhs.attachmentType),
-  loadOp(rhs.loadOp),
-  storeOp(rhs.storeOp),
-  clearFlags(rhs.clearFlags),
-  clearColor(rhs.clearColor),
-  slotID(rhs.slotID),
-  shaderStageFlags(rhs.shaderStageFlags) {}
-
-RasterView::RasterView(RasterView const& rhs, const allocator_type& alloc)
-: slotName(rhs.slotName, alloc),
-  slotName1(rhs.slotName1, alloc),
-  accessType(rhs.accessType),
-  attachmentType(rhs.attachmentType),
-  loadOp(rhs.loadOp),
-  storeOp(rhs.storeOp),
-  clearFlags(rhs.clearFlags),
-  clearColor(rhs.clearColor),
-  slotID(rhs.slotID),
-  shaderStageFlags(rhs.shaderStageFlags) {}
-
-ComputeView::ComputeView(const allocator_type& alloc) noexcept
-: name(alloc) {}
-
-ComputeView::ComputeView(ccstd::pmr::string nameIn, AccessType accessTypeIn, gfx::ClearFlagBit clearFlagsIn, ClearValueType clearValueTypeIn, ClearValue clearValueIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc) noexcept
-: name(std::move(nameIn), alloc),
-  accessType(accessTypeIn),
-  clearFlags(clearFlagsIn),
-  clearValueType(clearValueTypeIn),
-  clearValue(clearValueIn),
-  shaderStageFlags(shaderStageFlagsIn) {}
-
-ComputeView::ComputeView(ccstd::pmr::string nameIn, AccessType accessTypeIn, uint32_t planeIn, gfx::ClearFlagBit clearFlagsIn, ClearValueType clearValueTypeIn, ClearValue clearValueIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc) noexcept
-: name(std::move(nameIn), alloc),
-  accessType(accessTypeIn),
-  plane(planeIn),
-  clearFlags(clearFlagsIn),
-  clearValueType(clearValueTypeIn),
-  clearValue(clearValueIn),
-  shaderStageFlags(shaderStageFlagsIn) {}
-
-ComputeView::ComputeView(ComputeView&& rhs, const allocator_type& alloc)
-: name(std::move(rhs.name), alloc),
-  accessType(rhs.accessType),
-  plane(rhs.plane),
-  clearFlags(rhs.clearFlags),
-  clearValueType(rhs.clearValueType),
-  clearValue(rhs.clearValue),
-  shaderStageFlags(rhs.shaderStageFlags) {}
-
-ComputeView::ComputeView(ComputeView const& rhs, const allocator_type& alloc)
-: name(rhs.name, alloc),
-  accessType(rhs.accessType),
-  plane(rhs.plane),
-  clearFlags(rhs.clearFlags),
-  clearValueType(rhs.clearValueType),
-  clearValue(rhs.clearValue),
-  shaderStageFlags(rhs.shaderStageFlags) {}
-
 ResolvePair::ResolvePair(const allocator_type& alloc) noexcept
 : source(alloc),
   target(alloc) {}
