+++ native/cocos/renderer/pipeline/custom/RenderCommonTypes.h
@@ -32,7 +32,6 @@
 #include "cocos/base/Ptr.h"
 #include "cocos/base/std/container/map.h"
 #include "cocos/base/std/container/string.h"
-#include "cocos/base/std/hash/hash.h"
 #include "cocos/renderer/gfx-base/GFXDef-common.h"
 #include "cocos/renderer/pipeline/custom/RenderCommonFwd.h"
 #include "cocos/scene/Light.h"
@@ -197,114 +196,11 @@ enum class AccessType {
     WRITE,
 };
 
-struct RasterView {
-    using allocator_type = boost::container::pmr::polymorphic_allocator<char>;
-    allocator_type get_allocator() const noexcept { // NOLINT
-        return {slotName.get_allocator().resource()};
-    }
-
-    RasterView(const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept; // NOLINT
-    RasterView(ccstd::pmr::string slotNameIn, AccessType accessTypeIn, AttachmentType attachmentTypeIn, gfx::LoadOp loadOpIn, gfx::StoreOp storeOpIn, gfx::ClearFlagBit clearFlagsIn, gfx::Color clearColorIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept;
-    RasterView(ccstd::pmr::string slotNameIn, ccstd::pmr::string slotName1In, AccessType accessTypeIn, AttachmentType attachmentTypeIn, gfx::LoadOp loadOpIn, gfx::StoreOp storeOpIn, gfx::ClearFlagBit clearFlagsIn, gfx::Color clearColorIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept;
-    RasterView(RasterView&& rhs, const allocator_type& alloc);
-    RasterView(RasterView const& rhs, const allocator_type& alloc);
-
-    RasterView(RasterView&& rhs) noexcept = default;
-    RasterView(RasterView const& rhs) = delete;
-    RasterView& operator=(RasterView&& rhs) = default;
-    RasterView& operator=(RasterView const& rhs) = default;
-
-    ccstd::pmr::string slotName;
-    ccstd::pmr::string slotName1;
-    AccessType accessType{AccessType::WRITE};
-    AttachmentType attachmentType{AttachmentType::RENDER_TARGET};
-    gfx::LoadOp loadOp{gfx::LoadOp::LOAD};
-    gfx::StoreOp storeOp{gfx::StoreOp::STORE};
-    gfx::ClearFlagBit clearFlags{gfx::ClearFlagBit::ALL};
-    gfx::Color clearColor;
-    uint32_t slotID{0};
-    gfx::ShaderStageFlagBit shaderStageFlags{gfx::ShaderStageFlagBit::NONE};
-};
-
-inline bool operator==(const RasterView& lhs, const RasterView& rhs) noexcept {
-    return std::forward_as_tuple(lhs.slotName, lhs.slotName1, lhs.accessType, lhs.attachmentType, lhs.loadOp, lhs.storeOp, lhs.clearFlags, lhs.shaderStageFlags) ==
-           std::forward_as_tuple(rhs.slotName, rhs.slotName1, rhs.accessType, rhs.attachmentType, rhs.loadOp, rhs.storeOp, rhs.clearFlags, rhs.shaderStageFlags);
-}
-
-inline bool operator!=(const RasterView& lhs, const RasterView& rhs) noexcept {
-    return !(lhs == rhs);
-}
-
 enum class ClearValueType {
     NONE,
     FLOAT_TYPE,
     INT_TYPE,
 };
-
-struct ClearValue {
-    ClearValue() = default;
-    ClearValue(double xIn, double yIn, double zIn, double wIn) noexcept // NOLINT
-    : x(xIn),
-      y(yIn),
-      z(zIn),
-      w(wIn) {}
-
-    double x{0};
-    double y{0};
-    double z{0};
-    double w{0};
-};
-
-inline bool operator==(const ClearValue& lhs, const ClearValue& rhs) noexcept {
-    return std::forward_as_tuple(lhs.x, lhs.y, lhs.z, lhs.w) ==
-           std::forward_as_tuple(rhs.x, rhs.y, rhs.z, rhs.w);
-}
-
-inline bool operator!=(const ClearValue& lhs, const ClearValue& rhs) noexcept {
-    return !(lhs == rhs);
-}
-
-struct ComputeView {
-    using allocator_type = boost::container::pmr::polymorphic_allocator<char>;
-    allocator_type get_allocator() const noexcept { // NOLINT
-        return {name.get_allocator().resource()};
-    }
-
-    ComputeView(const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept; // NOLINT
-    ComputeView(ccstd::pmr::string nameIn, AccessType accessTypeIn, gfx::ClearFlagBit clearFlagsIn, ClearValueType clearValueTypeIn, ClearValue clearValueIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept;
-    ComputeView(ccstd::pmr::string nameIn, AccessType accessTypeIn, uint32_t planeIn, gfx::ClearFlagBit clearFlagsIn, ClearValueType clearValueTypeIn, ClearValue clearValueIn, gfx::ShaderStageFlagBit shaderStageFlagsIn, const allocator_type& alloc = boost::container::pmr::get_default_resource()) noexcept;
-    ComputeView(ComputeView&& rhs, const allocator_type& alloc);
-    ComputeView(ComputeView const& rhs, const allocator_type& alloc);
-
-    ComputeView(ComputeView&& rhs) noexcept = default;
-    ComputeView(ComputeView const& rhs) = delete;
-    ComputeView& operator=(ComputeView&& rhs) = default;
-    ComputeView& operator=(ComputeView const& rhs) = default;
-
-    bool isRead() const {
-        return accessType != AccessType::WRITE;
-    }
-    bool isWrite() const {
-        return accessType != AccessType::READ;
-    }
-
-    ccstd::pmr::string name;
-    AccessType accessType{AccessType::READ};
-    uint32_t plane{0};
-    gfx::ClearFlagBit clearFlags{gfx::ClearFlagBit::NONE};
-    ClearValueType clearValueType{ClearValueType::NONE};
-    ClearValue clearValue;
-    gfx::ShaderStageFlagBit shaderStageFlags{gfx::ShaderStageFlagBit::NONE};
-};
-
-inline bool operator==(const ComputeView& lhs, const ComputeView& rhs) noexcept {
-    return std::forward_as_tuple(lhs.name, lhs.accessType, lhs.plane, lhs.clearFlags, lhs.clearValueType, lhs.shaderStageFlags) ==
-           std::forward_as_tuple(rhs.name, rhs.accessType, rhs.plane, rhs.clearFlags, rhs.clearValueType, rhs.shaderStageFlags);
-}
-
-inline bool operator!=(const ComputeView& lhs, const ComputeView& rhs) noexcept {
-    return !(lhs == rhs);
-}
 
 struct LightInfo {
     LightInfo() = default;
@@ -520,42 +416,5 @@ struct PipelineStatistics {
 } // namespace render
 
 } // namespace cc
-
-namespace ccstd {
-
-inline hash_t hash<cc::render::RasterView>::operator()(const cc::render::RasterView& val) const noexcept {
-    hash_t seed = 0;
-    hash_combine(seed, val.slotName);
-    hash_combine(seed, val.slotName1);
-    hash_combine(seed, val.accessType);
-    hash_combine(seed, val.attachmentType);
-    hash_combine(seed, val.loadOp);
-    hash_combine(seed, val.storeOp);
-    hash_combine(seed, val.clearFlags);
-    hash_combine(seed, val.shaderStageFlags);
-    return seed;
-}
-
-inline hash_t hash<cc::render::ClearValue>::operator()(const cc::render::ClearValue& val) const noexcept {
-    hash_t seed = 0;
-    hash_combine(seed, val.x);
-    hash_combine(seed, val.y);
-    hash_combine(seed, val.z);
-    hash_combine(seed, val.w);
-    return seed;
-}
-
-inline hash_t hash<cc::render::ComputeView>::operator()(const cc::render::ComputeView& val) const noexcept {
-    hash_t seed = 0;
-    hash_combine(seed, val.name);
-    hash_combine(seed, val.accessType);
-    hash_combine(seed, val.plane);
-    hash_combine(seed, val.clearFlags);
-    hash_combine(seed, val.clearValueType);
-    hash_combine(seed, val.shaderStageFlags);
-    return seed;
-}
-
-} // namespace ccstd
 
 // clang-format on
