+++ native/cocos/renderer/pipeline/custom/RenderInterfaceTypes.h
@@ -228,14 +228,6 @@ public:
     virtual void addRenderTarget(const ccstd::string &name, gfx::LoadOp loadOp, gfx::StoreOp storeOp, const gfx::Color &color) = 0;
     virtual void addDepthStencil(const ccstd::string &name, gfx::LoadOp loadOp, gfx::StoreOp storeOp, float depth, uint8_t stencil, gfx::ClearFlagBit clearFlags) = 0;
     virtual void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) = 0;
-    /**
-     * @deprecated method will be removed in 3.9.0
-     */
-    virtual void addRasterView(const ccstd::string &name, const RasterView &view) = 0;
-    /**
-     * @deprecated method will be removed in 3.9.0
-     */
-    virtual void addComputeView(const ccstd::string &name, const ComputeView &view) = 0;
     virtual RenderQueueBuilder *addQueue(QueueHint hint, const ccstd::string &phaseName) = 0;
     virtual void setViewport(const gfx::Viewport &viewport) = 0;
     virtual void setVersion(const ccstd::string &name, uint64_t version) = 0;
@@ -335,10 +327,6 @@ public:
     virtual void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) = 0;
     virtual void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
     virtual void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
-    /**
-     * @deprecated method will be removed in 3.9.0
-     */
-    virtual void addComputeView(const ccstd::string &name, const ComputeView &view) = 0;
     virtual void setViewport(const gfx::Viewport &viewport) = 0;
     virtual RenderQueueBuilder *addQueue(QueueHint hint, const ccstd::string &phaseName) = 0;
     virtual bool getShowStatistics() const = 0;
@@ -429,10 +417,6 @@ public:
     virtual void addTexture(const ccstd::string &name, const ccstd::string &slotName, gfx::Sampler *sampler, uint32_t plane) = 0;
     virtual void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
     virtual void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
-    /**
-     * @deprecated method will be removed in 3.9.0
-     */
-    virtual void addComputeView(const ccstd::string &name, const ComputeView &view) = 0;
     virtual ComputeQueueBuilder *addQueue(const ccstd::string &phaseName) = 0;
     /**
      * @beta function signature might change
@@ -482,10 +466,6 @@ public:
     virtual void addStorageBuffer(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
     virtual void addStorageImage(const ccstd::string &name, AccessType accessType, const ccstd::string &slotName) = 0;
     virtual void addMaterialTexture(const ccstd::string &resourceName, gfx::ShaderStageFlagBit flags) = 0;
-    /**
-     * @deprecated method will be removed in 3.9.0
-     */
-    virtual void addComputeView(const ccstd::string &name, const ComputeView &view) = 0;
     virtual ComputeQueueBuilder *addQueue(const ccstd::string &phaseName) = 0;
     /**
      * @beta function signature might change
